Split consolidator in multiple classes.
This breaks the consolidator code into seperate classes for the
consolidation of array metadata, fragments, fragment metadata and
commits. It also moves the new classes to a separate directory. For
vacuuming, the code has been removed from the storage manager and into
the seperate consolidator classes.

---
TYPE: IMPROVEMENT
DESC: Split consolidator in multiple classes.
